+++ Sources/MathToMusic/Processors/HarmonicCombiner.cs
@@ -3,7 +3,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 
-namespace MathToMusic
+namespace MathToMusic.Processors
 {
     /// <summary>
     /// Combines multiple sequential tone tracks into harmonic chords
