+++ Tests/MathToMusic.Tests/HarmonicCombinerTests.cs
@@ -3,6 +3,7 @@ using MathToMusic.Contracts;
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using MathToMusic.Processors;
 
 namespace MathToMusic.Tests
 {
